make code follow std
