+++ web/src/api/image.ts
@@ -1,21 +1,1 @@
-import { getIdToken } from "../firebase/auth/lib";
-
-type URL = string;
-// todo: move this to api/internal/endpoints.ts
-const PFP_POST_ENDPOINT = "http://localhost:3000/pfp";
-
-export async function uploadImage(file: File): Promise<URL> {
-  console.log("sending image to server...");
-  const res = await fetch(`${PFP_POST_ENDPOINT}?token=${await getIdToken()}`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "image/png",
-    },
-    body: file,
-  });
-  if (res.status !== 201)
-    throw new Error(
-      `Unexpected status code: expected 201, but got ${res.status}`,
-    );
-  return await res.text();
-}
+export { uploadImage } from "./internal/fetch-func";
